algunos tests de integracion
