+++ src/components/factory/actions.js
@@ -56,12 +56,17 @@ export const moveMachine = position => {
   }
 }
 
+export const makeTick = () => {
+  return (dispatch, getState) => {
+    dispatch(doTick(findMachinesToTick(getState().factory)))
+  }
+}
 export const TICK = 'TICK'
 export const tick = () => {
   return (dispatch, getState) => {
     if (!getState().factory.timer) {
       let timer = setInterval(
-        () => dispatch(doTick(findMachinesToTick(getState().factory))),
+        () => dispatch(makeTick()),
         5000
       )
       dispatch(tickTimer(timer))
