+++ src/components/machines/direction.js
@@ -77,7 +77,7 @@ export const isPositionAtDirection = (position, direction, toPosition) => {
   return equals(positionAt(position, direction), toPosition)
 }
 
-const equals = (left, right) => {
+export const equals = (left, right) => {
   return left.x === right.x && left.y === right.y
 }
